Functional Search by ID
Updated the search.js file to somewhat properly search for an object by ID. Caught up a little on the saving to the database as we need functions from the history.js file to do that correctly. Some code may need to be removed or moved to the history.js file for the correct functionality.